feat(prettier): add prettierignore
